packages/sdk/pyproject.toml
[project]
name = "asset-sdk"
version = "0.1.0"
description = "Core SDK for the asset optimization pipeline"
requires-python = ">=3.12"
dependencies = [
    "asset-db",
    "pillow>=10.0",
    "rembg>=2.0",
    "reportlab>=4.0",
    "boto3>=1.34",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.0",
    "gspread>=6.0",
    "httpx>=0.27",
    "anyio>=4.4",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/asset_sdk"]
